Created seperate file for ASCII drawings.
