+++ HauntedHouse.py
@@ -5,47 +5,13 @@
 import sys
 from dice import diceRoll
 from helpers import *
+from asciiDrawings import *
 
 def main() :
     
     clearConsole(0)
     asciiHouse()
 
-
-def asciiHouse() :
-
-    print("\n")
-    print("                                               | |")
-    print("                                              =====")
-    print("                                              [   ]")
-    print("                    __________________________[   ]__________")
-    print("                   /                                         \\")
-    print("                  /                                           \\")
-    print("                 /                                             \\")
-    print("                 ------------------------------------------------")
-    print("                 |                                               |")
-    print("                 |   __________        _____         __________  |")
-    print("                 |   |         |       |   |         |        |  |")
-    print("                 |   |         |       |   |         |        |  |")
-    print("                 |   |         |       |   |         |        |  |")
-    print("                 |   |         |       =====         |        |  |")
-    print("                 |   ===========                     ==========  |")
-    print("                 |                                               |________________________")
-    print("                 |                                               |                         \\")
-    print("                 |   ________________                __________  |_________________________\\")
-    print("~~~~~~~~~~~~~~~~~|   |              |    ________    |        |  |                         |")
-    print("~~~~~~~~~~~~~~~~~|   |______________|    |      |    |________|  |                         |~~~")
-    print("~~~~~~~~         |   |              |    |      |    |        |  |                         |~~~")
-    print("~~~~~~~~         |   |              |    |      |    |        |  |                         |~~~")
-    print("~~~~~~~~         |   |              |    |    O |    |        |  |                         |~~~")
-    print("~~~~~~~~         |   |              |    |      |    ==========  |             O           |~~~")
-    print("~~~~~~~~         |   ================    |      |                |                         |~~~")
-    print("~~~~~~~~         |                       |      |                |                         |~~~")
-    print("~~~~~~~~         |                     [##########]              |_________________________|~~~")
-    print("'''''''''+++++++'''''''''''''''''''''[##############]''''''''''''+++++++++++++++++++++++++++''''")
-    print("''''''''''+++++++''''''''''''''''''+++++++++++++++++++''''''''''++++++++++++++++++++++++++++''''")
-    print("''''''''''''+++++++''''''''''''''''++++++++++++++++++''''''''''++++++++++++++++++++++++++++'''''")
-    print("\n\n")
 
 if __name__ == "__main__" :
     main()
