+++ asciiDrawings.py
@@ -1,0 +1,38 @@
+# Collection of ASCII drawings
+
+def asciiHouse() :
+    """Prints an ASCII graphic of the front of a house to the screen."""
+
+    print("\n")
+    print("                                               | |")
+    print("                                              =====")
+    print("                                              [   ]")
+    print("                    __________________________[   ]__________")
+    print("                   /                                         \\")
+    print("                  /                                           \\")
+    print("                 /                                             \\")
+    print("                 ------------------------------------------------")
+    print("                 |                                               |")
+    print("                 |   __________        _____         __________  |")
+    print("                 |   |         |       |   |         |        |  |")
+    print("                 |   |         |       |   |         |        |  |")
+    print("                 |   |         |       |   |         |        |  |")
+    print("                 |   |         |       =====         |        |  |")
+    print("                 |   ===========                     ==========  |")
+    print("                 |                                               |________________________")
+    print("                 |                                               |                         \\")
+    print("                 |   ________________                __________  |_________________________\\")
+    print("~~~~~~~~~~~~~~~~~|   |              |    ________    |        |  |                         |")
+    print("~~~~~~~~~~~~~~~~~|   |______________|    |      |    |________|  |                         |~~~")
+    print("~~~~~~~~         |   |              |    |      |    |        |  |                         |~~~")
+    print("~~~~~~~~         |   |              |    |      |    |        |  |                         |~~~")
+    print("~~~~~~~~         |   |              |    |    O |    |        |  |                         |~~~")
+    print("~~~~~~~~         |   |              |    |      |    ==========  |             O           |~~~")
+    print("~~~~~~~~         |   ================    |      |                |                         |~~~")
+    print("~~~~~~~~         |                       |      |                |                         |~~~")
+    print("~~~~~~~~         |                     [##########]              |_________________________|~~~")
+    print("'''''''''+++++++'''''''''''''''''''''[##############]''''''''''''+++++++++++++++++++++++++++''''")
+    print("''''''''''+++++++''''''''''''''''''+++++++++++++++++++''''''''''++++++++++++++++++++++++++++''''")
+    print("''''''''''''+++++++''''''''''''''''++++++++++++++++++''''''''''++++++++++++++++++++++++++++'''''")
+    print("\n\n")
+
